Finish prototype of Bloon/Dart/Monkeys
Currently monkey can detect nearby bloons and shoot, dart can go
forward, bloon can stay on track and most importantly, bloons can
be poped by darts and pierce count works wonderfully. User interface
and upgrades for monkey not tackled with.

Bloon mechanism will break if added damage: need some kind of
`List<Bloon> Bloon::popOneLayer()` to support that.

Gradle fixed such that it uses Java 9 to compile

+++ core/src/com/balloontd/game/Player.java
@@ -8,6 +8,6 @@ public class Player {
     public synchronized float getMoney() { return money; }
     public synchronized void setHP(float hp) { this.hp = hp; }
     public synchronized void setMoney(float money) { this.money = money; }
-    public synchronized void costHP(int cost_hp) { hp -= cost_hp; }
-    public synchronized void addMoney(int added_money) { money += added_money; }
+    public synchronized void costHP(float cost_hp) { hp -= cost_hp; }
+    public synchronized void addMoney(float added_money) { money += added_money; }
 }
